configure: added option to install additional CA certs
When pidgin is built with --with-system-ssl-certs, libpurple relies on
the system to have all required CA certs installed. But when GnuTLS is
used, the cert verification process can error when the issuing CA cert
of a CA cert is missing.

In order to navigate this issue, --with-ssl-certs has been added to
allow for the installing of potentially missing CAs. An installation
of pidgin where --with-system-ssl-certs is used is recommended to pass
this flag to autotools at configuration time.

This issue is seen with the Baltimore_CyberTrust_Root CA, which is
distributed by ca-certificates. The Baltimore certificate is issued by
the GTE_CyberTrust_Global_Root CA, and GTE_CyberTrust_Global_Root is
missing from ca-certificates.