configure: dropped required glib version to 2.20.0
It was initially thought the plugin would conform to 2.32.0 without any
sort of compatibility symbols. But with the plugin being right from the
upstream purple3 sources, which is using glib 2.20.0, this plugin has
already conformed to using 2.20.0. Drop the version to allow for people
who are stuck in the past to use the plugin.

Also worth noting, upstream purple3 is likely bumping the glib version
to 2.40.0. It is unknown if this plugin will continue to carry around
compatibility symbols, or if it will just simply require the newer
version of glib.