Update ThreadListQuery hash, fixes sync_sequence_id issues, thanks Eion
This hash represents a graphql query cached in the server. We were using
an old version of the hash that broke for some reason, possibly related
to the fact that fb messenger versions older than 5 are being
discontinued. For reference, the oldest parts of the plugin were written
when 4 was the latest version, most parts are updated to 33 or 38, and
the newest parts are based on 100 or so.

Thanks EionRobb for coming up with this fix so quickly.