Release 0.9.5 fixes "Failed to read thrift" errors
This isn't the fix. This is so that people see a recent commit that says
something is fixed and stop asking where the fix is, because they think
a commit from two weeks ago can't possibly be the fix. But it is.

Handling support for these cases is my least favorite part of this thing